+++ NEWS.md
@@ -1,17 +1,24 @@
-# qlcMatrix 0.9.6
+# qlcMatrix 0.9.8
 
+# qlcMatrix 0.9.7
+
+## changes
 * adding experimental distSparse function
 * adding option to pass arguments from sim.obs and sim.att to splitTable
 * adding option to pass arguments from sim.strings to splitStrings
 * changed internal gap symbol in pwMatrix and splitStrings to \u2043
 * adding dimRed() function for sparse dimensionality Reduction based on Cholesky()
 * adding normalized Laplacian norm 'normL' for cosSparse
-* adding commandline execs
 
 ## bugs
 * removed outdated links
 * replacing rBind/cBind with rbind/cbind as of R 3.2
 * corrected names of rKhatriRao for the complete symmetric case
+
+# qlcMatrix 0.9.6
+
+* adding commandline execs
+* importing docopt
 
 # qlcMatrix 0.9.5
 * finally adding a NEWS list